requirements.txt
pyFirmata ~= 1.1.0
keras ~= 2.11.0
tensorflow ~= 2.11.0
numpy ~= 1.24.0
Pillow ~= 9.4.0
requests ~= 2.28.2
beautifulsoup4 ~= 4.11.1
opencv-python ~= 4.7.0.68
joblib ~= 1.2.0
